+++ index.html
@@ -229,7 +229,7 @@
   </template>
 
   <!-- <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script> //подключение lodash -->
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
